requirements.txt
# PDF Forensics Detection System - Requirements
# EAS 510 Assignment 3

# Core ML and Data Processing
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Image Processing
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# PDF Generation Libraries
reportlab>=4.0.0
fpdf>=1.7.0
python-docx>=1.1.0
docx2pdf>=0.1.8

# System and Utilities
requests>=2.31.0
tqdm>=4.65.0

# Development and Testing
pytest>=7.4.0
jupyter>=1.0.0